refactor and clean up Chains component
