--- src/components/molecules/UserPreferences/Chain.module.css
@@ -1,48 +1,0 @@
-.chain {
-  margin-right: calc(var(--spacer) / 3);
-}
-
-.titleGroup {
-  font-size: var(--font-size-small);
-  margin-bottom: calc(var(--spacer) / 6);
-  margin-top: calc(var(--spacer) / 2);
-  color: var(--color-secondary);
-}
-
-.chains {
-  composes: box from '../../atoms/Box.module.css';
-  box-shadow: none;
-  padding: 0;
-}
-
-.radioWrap {
-  composes: radioWrap from '../../atoms/Input/InputElement.module.css';
-  padding: calc(var(--spacer) / 6) calc(var(--spacer) / 3);
-  border-bottom: 1px solid var(--border-color);
-}
-
-.radioWrap:last-child {
-  border-bottom: none;
-}
-
-.input {
-  composes: checkbox from '../../atoms/Input/InputElement.module.css';
-  vertical-align: baseline;
-  margin-right: calc(var(--spacer) / 3);
-  margin-top: 0;
-}
-
-.radioLabel {
-  composes: radioLabel from '../../atoms/Input/InputElement.module.css';
-  font-weight: var(--font-weight-base);
-  display: flex;
-  align-items: center;
-  margin: 0;
-  padding: 0;
-  width: 100%;
-}
-
-.input:checked + span {
-  color: var(--font-color-text);
-  font-weight: var(--font-weight-bold);
-}
