+++ src/components/molecules/UserPreferences/Chains/index.module.css
@@ -1,0 +1,3 @@
+.chain {
+  margin-right: calc(var(--spacer) / 3);
+}
